hpcdata: fix callers view creation, especially for compact inlined procedure names.
The current hpctoolkit-prof branch will compact all the same procedures
into one single procedure within the procedure table.
However, for the inlined procedures, the line scope (or call site) has
an empty procedure name. This causes all inlined procedures have exactly
the same flat ID for its line scope.
This line scope was used by callers view builder to check whether a
call site is in the same path as the existing tree or not. Since all
inlined procedures have the same line scope, the builder merges them
into the same path.

The solution is to use the procedure scope of the call site instead of
the line scope to check if they are in the same path or not.