hpcviewer Bug fix: merging two flat trees into one database
"Node Id" for flat tree is not unique. We reuse the ID from file table
nd load table for files and load module respectively.

However, the same ID can be reused for CCT node. 
For instance if file A has ID "2", it is possible a node foo has ID "2".

Therefore, we shouldn't assume node ID will be unique in flat tree. 

In the TreeSimilarity class, we "cache" the list of sorted children with
the goal to avoid resorting the same thing. However, the list is stored
in a hashmap where the key is the node ID of the parent. Since this
doesn't work for merging flat tree, we shouldn't use cache mechanism.
Moreover, it's better to recompute and save memory instead of saving
computation and waste the memory.