hpcviewer: accommodate the lack of info on metric value for the exclusive metric of the root. hpcprof doesn't give us the info of the exclusive metric of the root, but now it provides the info of the metric type and its partner for thread-level metric.
So if the exclusive metric has no value for the root, we just "peek" the
value of its partner (which is inclusive).
Right now we just assume all exclusive metric has no value for the root
(unlike the traditional metrics). In the future we should make hpcprof
consistent for both traditional metrics and thread-level metrics.