hpcdata and hpcviewer: remove direct dependency to library jar file. Instead we now depend on the plugin of the library instead of the jar files.
Pros: it is easier to merge with the original repository and the effect
of the updates can be seen automatically. While using jar, we have to
merge first, export into jar, and then import it into hpcviewer/hpcdata.
Now we just do one step instead of four.

Cons: hpcviewer and hpcdata now requires hpcviewer-externals repository
to be in the same Eclipse workspace.