hpcviewer: bug fix when merging databases with a derived metric in the second database.
The preference is to not including a derived metric into the merged
database since modifying the formulae can be tricky.
For instance, assuming the second database contains a derived metric
with the formula "$2+$3". And assuming the metric index 2 is changed
into "10" and metric index 3 into "11" (due to merge with the first
database).
To make the formula correct, index 2 has to be replaced with 10 and
index 3 into 11.