hpcviewer: preparation for release 5.4.1
set the Filter plugin loaded as default. It's ready to be released.

Fix special case for inline procedure: if the procedure name is "-" or
"<inline> then we treated it as guard alien.

Fix naming for Procedure scope. Instead of generating the name of the
procedure dynamically, we created once during the object creation. this
should improve the performance since Java VM is bad with String.