hpcdata: bug fix when computing exclusive metrics when the filter mode is on.
Originally (dated 2002), hpcview distinguish between a valid metric
value with non-existence metric value by using number -1. If a metric
has this value, then the scope has no value.

However, this complicates when we want to recompute metric values such
as in filter. 
For instance  if we filter (exclude) node A, then the inclusive cost of
the children of node A becomes the exclusive cost of the parent of node
A. If the children of node A has no value (i.e. the value is "-1", then
aggregating to the exclusive cost of the parent of node A will decrement
the cost of the parent!
Hence we always need to check if the children has value or not before
aggregating to the parent. It is not a proper solution, but so far it
works.