hpcfilter: minor fix for toggle menu in "show the filter view"
Handling toggle menu/command in Eclipse is either trickier than I
thought, or it's buggy.
At the moment, it's simpler to use "push" menu intead of "toggle" menu.
If the filter view is already created, clicking on "show the filter
view" will activate the view instead of hiding it (which is the toggle
way)