hpcfilter: using only one switch to enable/disable filter patterns.
Using two switches have the advantages of turning on/off filter if we
need to modify a couple of filter patterns. 
For instance, we create filter A, B and C and then turn the filter on to
broadcast filter message to all windows. This saves us time instead of
using one switch where we broadcast filter message everytime we create a
single filter.

However this has major disadvantage that it can confuse users. If the
global switch is off, then a user enable a filter, he/she expects to
have immediate effect. However, hpcviewer will show no modification
since the the global switch is turned off.

To reconciliate this, it is better to have one switch with a dialog box
to show the property of filter. Users can modify as many as they want,
and then click "ok" to see the effect. 
With this approach, the "global switch" is always on. This may also
confuse users if he/she forgets that filter is always enabled.