hpctraceviewer: preparation for 5.4.1 release
- minor bug fix on summary view. this is mainly an Eclipse bug. Somehow
Eclipse reports that a certain pixel when convert to rgb color it
becomes white (255, 255, 255). to minimize this effect, we need as much
as possible to "round up" the height of other colors. For instance for
color blue, if the height is 5.1 pixels, then we consider it to 6. Since
everything is round up, there will no place for white color (in theory).

- make open local database to be consistent. If we have two accesses to
open local db: one during the launch of the app, and the other one from
the file menu. To reduce code redundancy, we should merge the two codes,
and make it the same class and the same treatment as the remote
database action.