hpcviewer: fix filtering
Filter can be both useful and harmful.
It is harmful when the filter remove the entire tree or the subtree of a
fake procedure (such as <thread root> or <program root>).
Fake procedures is shown in CCT but not in callers tree or flat tree.
Hence, if the children of a fake procedure are omitted, then the
exclusive cost of the children will be attributed to the fake
procedure.

Since fake procedures are not shown in callers tree and flat tree, we
can see discrepancy and incomplete information, which results misleading
information.

What we can do at the moment is to warn people to be careful in using
Filter. If we detect it has changed fake procedures, we show a
notification.