hpcviewer: fixing dynamic flat tree when we want to merge databases
When we merge two flat trees, we need to force each database to create
the flat tree if it isn't created yet (it's created dynamically).
On the other hand, the flat view (within the viewer) needs to detect if
the flat tree has been created without its knowledge or not. If this is
the case (like in merging case) we need to reset the input to take into
account the new created flat tree.

Another issue with the mix between filter and merging, right now, if the
merging only considers the current condition of the tree. If the tree is
filtered and then merged with another tree, the result will be a
filtered merge tree.
If then we decide to cancel the filter, the merged tree still shows the
filtered one. This is a known bug, but I don't know how to handle this
properly.