hpcviewer: revert to the original pre-computed callers tree and flat tree.
One of the issues with dynamic tree is we need to deal with
"derived incremental" metrics (they are metrics generated by hpcprof for
statistic outputs such as mean, std dev and covariance). These metrics
comprise of two states: combined and finalized. Unfortunately, the
original design of cct is to finalize every metric values during the
post-processing of xml parser. 
Once a metric is "finalized", its value cannot be used anymore for
computing metrics for callers tree and flat tree (since they are not in
"raw" form anymore).

What we need to do is to conserve the original metric values before
finalization, or never finalize it at all. We only finalize it for the
display only.
The first option requires more storage which is not appropriate. The
second option requires more computation but the number of computation is
the same order of the number of visible metrics on the table.