hpcdata and hpcviewer: draft for dynamic metric
Before we create flat tree and callers tree dynamically, we should
provide a dynamic metric computation. For "traditional" metric, it's
easy, there is nothing to be computed. For "derived-incremental"
metrics, we have to be more careful since these metrics have formula to
compute their "final" values.

Originally, we provide two metrics for each scope: a combine metric and
a finalized metric. the latter is pre-computed after processing CCT. We
deliberately store the combine metrics since callers tree is created
dynamically, and it requires CCT's combine metrics to compute its
finalized values.

In this commit, we compute the finalized metrics on the fly, and we do
not need to store two metrics for each scope.
A major disadvantage is that we requires more computation power since we
always compute the finalize version every time a scope is visible.