hpcviewer: fixing bugs when combining between dynamic tree creation and filtering.
When we apply filter, we need to refresh the all trees (cct, callers and
flat tree). However, this had conflict with the dynamic tree approach
since the later only wants to create the tree if the children are not
created. Forcing to refresh the tree is the only easy option at the
moment. We need to find a better approach.