hpcviewer: fix mixed up derived metrics and filter.
A bug occurs when we add a derived metric without activating callers
view and flat view. Then, if we apply filtering, the derived metric(s)
disappear since the viewer has to reparse the database.

The solution is we need to conserve the user-defined metrics before
applying filter, and then restore these metrics once the filter
terminates.