hpcviewer: [draft] Separate between metric and format style
We need to distinguish between metric data and metric presentation. For
instance, in some cases we want to display metric with scientific format
but sometimes we want to display with normal format (usually if the
metric value is not big). 
Currently, metric class is tightly coupled with its presentation, make
it hard to change the format, and even worse, the annotation format
(such as percentage).