hpcviewer: draft separation between format style and metric
Originally, we have a tight relationship between format style and
metrics. This causes a headache when we want to change format or change
annotation style as we need to make surgery to many parts of the code
just to update a metric.

Minor bug fix  for resetting experiment to instantiate thread level
metric class after filtering (once we filter, we need to read again the
database and we lose the information of thread-level metrics)
This commit is a draft for separating between metrics and their
format style. We need to separate as well with the annotation so that we
just need to compute the percentage in one place only.