hpcviewer: fix tree merge feature with dynamic tree and metric
The new database format supports dynamic metric and annotation. The
metric is not loaded and the percentage is not computed unless the scope
is "displayed" on the view.

For the merging, we need to force to load the metrics and compute the
percentage. 

TODO: this works fine with merging CCT but somehow for merging flat
trees the percentage is computed.