hpcviewer: temporary fix for not filtering a merged database
Usually, when a user activates a filter, a notification is sent to all
views to filter their tree. However, for a merged database is a bit
different. First, they don't have physical database, so they are not
"real" database. Second, I am not sure the semantic of "filtering a
merged database". Does it mean we filter the already merged database, or
we merge two filtered database ? 

Intuitively filtering a merged database means to merge two filtered
database, not filtering a merged database.

At the moment, let's not to filter a merged database until we know
what we want.