hpcviewer: avoid to render flat/callers view when we are closing
One of the problems with Eclipse view notification it that when we are
closing a view, first it forces the view to be visible, and then send
close notification. Hence, if a flat tree is never constructed, since
the flat view becomes visible, we are building a flat although we have
to destroy it later on.
This causes closing a database seems slow especially for large
databases.

To avoid this, we should mark that a database is about to be removed,
and then tell Eclipse to send notification to close the views associated
with the database.