hpctraceviewer: fix bug when displaying empty percentage in summary view
The root cause of the bug is we store the list of pixels from the
original image (main view), then draw from the list to the summary view.
When a cursor moves over on a summary view, we copy the image of the
pixel pointed by the cursor, then covert this image into a pixel.

However, it turns out, somehow, that the copied image may not be the
same as the original summary view image (which is Eclipse bug?).
The solution is to force to use the original summary view image, and
then covert an image where the cursor is located, into a pixel.

This approach is more direct and works well on Mac.