hpctraceviewer: temporary fix of data races in depth view
The cause of the races is we use background job to paint the depth view.
The problem with background job is that once we define the job, we'll
return immediately and we don't know when it will start or terminate.
If another event occurs while we are still painting the depth view (or
the job is still in the queue) it will cause data races.

A temporary solution is to avoid to use background job for depth view.
We need to redesign the painting so that if an event occurs, it can
terminate the current job.