hpctraceviewer: job cancelation when depth view has data races
One of data races occur when user change the bounds of the current
process position very quickly while the viewer is still painting the
depth view.
In this case, it may happen that the current process is null and it will
cause the data gathering phase terminates earlier that it should be. For
this, another phase (painting phase) has to notified about early
termination. There are some options here, but it seems the easiest way
is to announce job cancelation or termination.

There should be a way for task dependency in Eclipse jobs and Java
threads and to express task cancelation without using exception (which
may cause chaos since it is global).