hpctraceviewer: guard the painting from accessing out of bound array
We still have a data race in main view (main painting). The data race
occurs when we zoom in or zoom out very-very quickly so that the
global variable that stores the list of process timeline (in
ProcessTimeineService class) is not consistent anymore. 

Temporary fix:
Instead of removing the global variable (share variables are always
problematic), we just check if the list of process timeline is big
enough for the current usage or not. If it is the case (although the
list not relevant anymore due to changes from other event) than we
proceed.
Otherwise we cancel the event job and give up. The last event should
have the correct one.
Note: there's no guarantee of FIFO in Eclipse UI events, i.e. if the
latest event in the list IS really the latest event that the user does.