hpcviewer: fix exclusive cost attribution for inline procedures
This occurs only with the flat view:
The original algorithm is to attribute exclusive cost of an inline
procedure to the file of the closest enclosing non-inline procedure.
I think we decided this because we want to assume inline procedures just
like loops.

The fix is to attribute the cost to the file where the inline procedure
is defined. This is consistent with the principle that inline procedure
is the same as normal procedures.