hpctraceviewer: fix problem with early painting termination
The problem occurs mainly on Linux/GTK but it doesn't occur on nancy
(and vnc on nancy). 
The root cause of this is hpctraceviewer terminates painting when we
finish collecting data trace. 

I believe On Mac and some Linux/GTK this doesn't occur if UI threads is
quick enough and it terminates when data collection also terminates. If
the UI thread is slow, then it thinks painting is done when in fact only
data collection is done.

The commit in cef6129 removes one of three conditions to prevent
early.

Another bug fix: data collector notifies paint threads if there is empty
trace in the list.