hpctraceviewer: ignore a not-critical exception in depth time canvas
The exception occurs when another thread access an image buffer that is
already initialized (disposed). In this case, either the thread returns
a unwanted value, or throws an exception. We prefer an exception since
it simplifies the code and a more natural way to do (the event thread
gives up the painting and let the existing event to finish the
painting).
However, extra exception message may scare users. So we just remove this
exception to avoid confusion.