hpctraceviewer bug fix: remove right click on mini map
On minimap, there is no need for right click or save the image. the
minimap has to tell the parent not to create context menu. 
Also, we need to distinguish between left click and right click on
minimap.

minor enhancement: remove unnecessary exception messages when a
processing task is replaced by another task. For instance, when a user
opening a large data, and then decided to open a small data, he/she
doesn't need to wait nor to cancel the current process. It is sufficient
just to open another data, and the old one should terminate
automatically (due to exception).