hpcdata: bug fix for reseting the root value of derived metric
Every time we change derived metrics, we need to force to update the
value of the root.
For performance reason, we cache the value of the root since it will be
needed every time we want to compute a percentage. The consequence of
this is that we should update the root value is the formula is changed.