hpcviewer bug fix: smarter way to get metric value
Originally, to take into account the "new" sparse metric format, we
decided to not to store "finalized" metric in the memory. All metric
values will be computed on the fly.
Hence, every time we ask a metric value, we query a metric to gives us a
value for a given scope:

value <- metric( scope )

However, this will cause a derived metric to retrieve the "original"
metric value of a metric instead of the "finalized" value. To fix this,
the work-around is the MetricVarMap to return original value if the
asking metric is the same as the metric to be queried. Otherwise, we
just return the finalized value.

Other changes: fixing warnings.