hpctraceviewer: temporarily adding refresh button
Due to bugs in Eclipse (I think), we may need to refresh and repaint
canvas to ensure that the view is populated.
On Mac OS platform, if we display summary view and then resize the
window, and go back to the depth view, the latter will display blanc
image.

This only happens on Mac, I am not sure the root cause of this bug. 
Hence, temporarily, we just add a toolbar on the depth view to refresh
the data and repaint....