rename to release 5.4.3 (based on mars)
this branch uses Eclipse 3.x compatibility layer which causes minor
issues (I think mainly on Linux) such as perspective layout. On ppc
little endian is a bit critical as the font and image are not well
placed, and the GC on the click of the tree doesn't work

We need to move to pure Eclipse 4 to avoid hack here and there.