hpctraceviewer: minor fix for the label format
Instead of using manual arithmetic operation, we delegate the formatting
to Java DecimalFormat, and let it compute the round-up (or down).

Attention: somehow, the current process position in the cross hair is
not exactly correct if screen height resolution is much smaller the the
number of processes to display. It looks like the position is off 1
rank. need to investigate further.