hpctraceviewer bug fix: consistent current process information
Due to inconsistency floating-point rounding (sometime we use ceiling,
sometime math rounding, sometimes integer rounding), we can encounter
inconsistency of the current selected process.

A more reliable way to display the current selected process is by using
the information of the current process in the depth view. However, to
know this info, we have to wait until the information is available (as
we use lazy rendering and lazy info gathering).