fix hpctraceviewer bugs: two bugs related to remote server
bug 1: the server always closes when we open the second database. The
root cause is we want to clean up the memory and accidently we close the
database as well. Solution: not closing the database while cleaning up
memory.

bug 2: since we use Depth view's ProcessTimeline, we depend on the
information of this class. However, for remote database, there is no
information about the process number, but the information of the line
number is available. Somehow using this info also works for remote
database.