hpctraceviewer bug fix: exception when resizing is faster then rendering
A corner case when we resize the window, but still in the process of
rendering, and user perform an action (such as hovering on Summary view
to see the statistics), Eclipse will give incorrect cursor position
which doesn't match to the canvas boundary.

In this case, instead of throwing an exception, we just place a guard
not to perform further action.