hpcviewer mars: hack due to Eclipse bug on Linux ppc little endian
On this platform, Eclipse gives incorrect position of the selected tree
item. When a user clicks an item (procedure name for instance) in a
tree, Eclipse sends notification the coordinate and size of the item.
However, the coordinate y and the height is incorrect. Is we want to
know if the click is part of the icon or the line number or the text, we
should compare only with the coordinate x. 

Note:
This bug only occurs on Linux ppc little endian. I don't see it in other
major platforms: linux x86, win and mac OS. It is perhaps not the
platform itself, but the window manager