hpcviewer mars: make sure the tree column has priority to have spaces
Although this problem mainly occurs in Linux environment, the fix us
general enough and can be applied for all platforms as well.
We need to use TableColumnLayout to give hint to Eclipse which column
needs to resize. We still not in the control of column size, but it is
better than not to give hint at all and let Eclipse guesses.

One major issue with this approach is we need to have a composite that
wraps the table. This composite, somehow, can cause the table's
horizontal scrollbar partially hidden. In some window sizes, this
doesn't occur. Not sure why,