hpcviewer mars bug fix: remove autoresize column when changing font
Eclipse mars caches the font of all tree items. However, changing the
font such as increasing/decreasing font sizes require to dispose the
current font and create a new font with the new size (there is no way to
update font without creating a new font). 
Since the old font has been disposed, resizing columns (using
tablecolumnlayout class) will cause Eclipse swt to compute the size of
the column with the old font which is stored in the cache. This causes
an exception.

To fix this, either not to resize automatically columns, or not to
change the font. Propagating the new font to all tree items is just too
painful.