hpcdata: Fix flat view to include inline functions (aliens)
In experiment.xml, inline macros are identified as "fake" procedure
where the value of attribute "f" is "1". Because of this, the flat view
visitor decides not to convert any inline macros into flat tree.

This commit allows any "fake" procedure as long as it isn't "root" or
"partial callpath" (or something like that) we will try to convert it
into flat tree. 
However, since inline macros are fake procedure AND the flat id is bogus
it causes an issue in the flat tree. The problem is that the same flat
id of an inline macro also used by many other nodes. For instance, an
inline macro X with flat id 1000 defined in file F, is also used by
other macros defined in file Y. To solve this, hpcdata needs to generate
its own flat id to differentiate inline macros defined in different
files.