hpcviewer: fix flat view visitor to generate inline macros
There are 2 types of inlines (aliens): inline functions which contain
callsite and a static inline procedure; and inline macros which contain
a single inline procedure. The former is treated like a normal
callsites, while the later is treated like a loop: embedded to the
parent.

As inline macro is embedded to the parent (or caller), in flat view it
doesn't need to be attached to its own file or load module.