Fix performance problem when sorting a column
I think the root causes are:

1. Using virtual table but not applying the lazy approach. The table is
set to be virtual, but doesn't use the correct lazy approach. The
lazyContentProvider should handle data on-demand thus, make the loading
faster.

2. Explicit refresh. In Eclipse Neon, it fixes bugs, so explicit refresh
after soring is not needed. Refreshing table takes time.

3. Slow interface to GTK. Displaying the content of table requires
setting the font, color foreground and background. Surprisingly, with
GTK, it causes around 50% of the time.

This commit fixes problem 1 and 2. For problem 3, we need to avoid
setting font and colors.