Special handling of lazy content provider for dynamic content
Dynamic content like callers view (where the tree is created
dynamically) needs a special handling with lazy content provider. In
dynamic tree, we don't know yet if a scope has children or not.
The lazy content provider needs to depend on polymorphism to get the
list of children, and the number of children (consequently if a scope
has a child or not).