Rename the title of the views
Rename from "Calling context view" to "Top-down view"
Rename from "Callers view" to "Bottom-up view"

Other minor update:
When we quit the viewer, we have to close the bottom-up and the flat
view first since they are created dynamically.
If we close the Top-down view first, Eclipse will activate the Bottom-up
view, and then close it. If the bottom-up tree is not created, when the
view is "activated", it builds the tree first, and then close it. This
may take significant time.