Fix thread view by using the sum instead of the average
Originally in thread view, we display the average of selected threads.
It is more intuitive to show the sum of all selected threads instead of
the average.

Minor fix: avoid selecting a row automatically when we display a view
for the first time. Using lazy content provider means the data is not
ready yet although the view is started to be visible.