hpcviewer: fix styler in the viewer to reflect change of font
The bug only happens in Scope column of the table.
Using the default StyledString.Styler attribute for tree's font style
doesn't reflect change of font (via increase/decrease font size, or via
preferences). 

To take into account change of font, we need to have our own Styler
attribute that depends on the viewer's font. Hence, whenever users
change the font, the change will reflect on the tree as well.

Note: for metric columns, there is no problem with this because we don't use the default font in StyledString.Styler.