hpcviewer: some fixes specifically for Windows platform
- Avoid using multi-layer dialog box. A bug on Windows, causes locking a
dialog during the starting hpctraceviewer for the first time. This bug
is fixed on Eclipse mac, but not on Windows. A workaround is to use
directory dialog directly without intermediate from the open database
dialog.

- standardize vm arguments between hpcviewer and hpctraceviewer. Avoid
using XX:MaxPermSize since this is not supported anymore on Java 8 and
newer. (Ecliose neon doesn't support Java 7 and older anyway).

- Clean up the code