Add required Java version in osgi parameter
Using bash shell to detect Java version is problematic. 
For instance, OpenJDK for Java 1.8 and older, the version is "1.8".
But for Java 10, the version is "10.x"

The most reliable way to detect Java version is to let Eclipse OSGI to
check the Java runtime version, and issue an error message if the
required version is not satisfied.