hpctraceviewer: limiting the number of created handles
1. Refine hpctraceviewer to generate only a certain number of color and
imgaes handles as some OS (like Windows) have limitation of the number
of handles.
If we have more procedures than the number of maximum, we randomly chose
existing colors-image to be reused.

2. Avoid generating colors/images on the fly.
On some platforms, we cannot generate colors/images on the fly as it may
break the concurrency. It's safer to generate colors/images during the
initialization.