Update launch script to take into account machine configuration
On some systems, it is possible to share home directory between
different machine configuration. For instance, user can have the same
home directory between x86_64 and ppcle. 
This causes a problem since Eclipse tries to store data and libraries
inside the home configuration directory.

To avoid this problem, we add a new layer to take into consideration the
machine configuration.