Add draft hpcread.py script to read experiment.xml
The script is useful to display the content of experiment.xml on ARM platforms where hpcviewer is not available.
Right now, it just visualizes procedure frame nodes, its file and the lenght of callpath.

It can be also useful for automatic testing.

The script is a modified version from LLNL's Hatchet tool.